Modified the draw-board function so that it reads of a linked list.
In order to avoid double keeping a lot of information about
which paths are allowed and which player is where I made a simple
linked list data structure which also contains the screen
coordinates for drawing. It also keeps which nodes you are
allowed to move to and whether or not it is a stone node or the
final (goal) node.